Combined code to one file and added `arazzo` key
