+++ custom-plugin-rules/default-enum-match/default-enum-plugin.js
@@ -1,12 +1,32 @@
-const DefaultEnumMatch = require('./default-enum-match');
-
 module.exports = function myRulesPlugin() {
   return {
-    id: 'openapi-default-enum',
+    id: 'default-enum',
     rules: {
       oas3: {
         'default-enum-match': DefaultEnumMatch,
       },
+      arazzo: {
+        'default-enum-match': DefaultEnum,
+      },
     },
   };
 };
+
+function DefaultEnumMatch() {
+  return {
+    Schema: {
+      enter(schema, ctx) {
+        // If enum and default are defined, default must be one of the enum values
+        if (
+          schema.enum &&
+          typeof schema.default != "undefined" &&
+          !schema.enum.includes(schema.default)
+        ) {
+          ctx.report({
+            message: `The default value must be one of the enum values.`,
+          });
+        }
+      },
+    },
+  };
+}
